pyproject.toml
[tool.poetry]
name = "chainlit-claude"
version = "0.1.0"
description = ""
authors = ["AZ <azuev@outlook.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
chainlit = "^1.1.402"
langchain-anthropic = "^0.1.22"
python-dotenv = "^1.0.1"
langchain = "^0.2.12"
firecrawl = "^0.0.20"
slugify = "^0.0.1"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
